requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==3.2.15
django-allauth==0.51.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
oauthlib==3.2.1
pathspec==0.10.1
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23
